refactor: topping type repository 파일 경로 수정
