+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/persistence/ToppingTypeRepository.kt
@@ -1,4 +1,4 @@
-package bingsoochef.bingsoochef.toppping
+package bingsoochef.bingsoochef.toppping.persistence
 
 import bingsoochef.bingsoochef.toppping.domain.ToppingType
 import org.springframework.data.jpa.repository.JpaRepository
